refactor: AppViewModel 을 Create, Edit 으로 분리
