+++ ProjectManager/ProjectManager/View/DetailView/DetailView.swift
